Grouped components into shared directory
