+++ frontend/foodieland-app/components/shared/Promosection.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Logo from "@/components/Logo";
+import Logo from "@/components/shared/Logo";
 import { useEffect, useState } from "react";
 
 export default function Promosection() {
